feat: add command-line interface
